refactor(slip): extract slip footer to separate component
